Refactored project structure II.
