+++ src/example/java/net/igsoft/sdi/ParametrizedCreatorExample.java
@@ -4,7 +4,7 @@ import java.io.File;
 
 import net.igsoft.sdi.creator.AutoCreator;
 import net.igsoft.sdi.creator.CreatorBase;
-import net.igsoft.sdi.internal.InstanceProvider;
+import net.igsoft.sdi.engine.InstanceProvider;
 import net.igsoft.sdi.parameter.ParameterBase;
 
 public class ParametrizedCreatorExample {
