+++ src/example/scala/net/igsoft/sdi/ParametrizedCreatorScalaExample.scala
@@ -3,7 +3,7 @@ package net.igsoft.sdi
 import java.io.File
 
 import net.igsoft.sdi.creator.{AutoCreator, CreatorBase}
-import net.igsoft.sdi.internal.InstanceProvider
+import net.igsoft.sdi.engine.InstanceProvider
 import net.igsoft.sdi.parameter.ParameterBase
 
 object ParametrizedCreatorScalaExample {
