+++ src/example/scala/net/igsoft/sdi/SimpleCreatorScalaExample.scala
@@ -1,7 +1,7 @@
 package net.igsoft.sdi
 
 import net.igsoft.sdi.creator.CreatorBase
-import net.igsoft.sdi.internal.InstanceProvider
+import net.igsoft.sdi.engine.InstanceProvider
 import net.igsoft.sdi.parameter.LaunchType
 
 object SimpleCreatorScalaExample { // tag::config[]
