+++ src/main/java/net/igsoft/sdi/engine/KeyGenerator.java
@@ -1,4 +1,4 @@
-package net.igsoft.sdi.internal;
+package net.igsoft.sdi.engine;
 
 public interface KeyGenerator {
     String generate(Class<?> clazz, String serializedParameters);
