+++ src/main/java/net/igsoft/sdi/engine/ManageableState.java
@@ -1,4 +1,4 @@
-package net.igsoft.sdi.internal;
+package net.igsoft.sdi.engine;
 
 public enum ManageableState {
     CREATED,
