+++ src/test/java/net/igsoft/sdi/testclasses/ACreator.java
@@ -1,7 +1,7 @@
 package net.igsoft.sdi.testclasses;
 
 import net.igsoft.sdi.creator.CreatorBase;
-import net.igsoft.sdi.internal.InstanceProvider;
+import net.igsoft.sdi.engine.InstanceProvider;
 import net.igsoft.sdi.parameter.LaunchType;
 
 public class ACreator extends CreatorBase<A, LaunchType> {
