+++ src/test/java/net/igsoft/sdi/testclasses/RWithDefaultCreatorsDoubledCreator.java
@@ -6,7 +6,7 @@ import com.google.common.collect.Lists;
 
 import net.igsoft.sdi.creator.AutoCreator;
 import net.igsoft.sdi.creator.CreatorBase;
-import net.igsoft.sdi.internal.InstanceProvider;
+import net.igsoft.sdi.engine.InstanceProvider;
 import net.igsoft.sdi.parameter.LaunchType;
 
 public class RWithDefaultCreatorsDoubledCreator extends CreatorBase<R, LaunchType> {
